ml/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==2.1.0
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
Pillow==10.1.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0


